requirements.txt
TestLink-API-Python-client
PyQt5>=5.14.2
Pillow>=7.1.1
